Upgrade to agent v6 & remove everything but checks
Configuration can easily be managed at the Kubernetes template level. No need
to include it in the image. Also, the agent now has support for including
secrets, so the custom entrypoint wrapper is also rendered unnecessary.

Checks are kept because they're a bit too complex to nicely insert from
ConfigMaps + they may require additional packages to be installed for them to
work correctly. E.g. the `conntrack-tools` used for the `conntrack` check.