Fix conntrack check
There was a couple of issues here.
One, `iteritems` was left over from when the returned value was a dictionary,
not a list.
Two, the result of running `conntrack.split('\n')` also includes a `None`, for
some reason, in real usage. Probably something to so with whitespace. Could've
appended `if line` to it to fix the issue, but instead I made the `line_regex`
more lenient (not proven to be needed, but just to be safe) and add the `if`
check for regex matching instead. This has the same outcome but should work
better in other situations with unexpected whitespace.