Fix subprocess execution in conntrack check
According to [DataDog documentation][1], direct `subprocess` module usage isn't
supported in v6 of the agent:

> Using the `subprocess` and `multiprocessing` modules provided by the python
standard library is not supported, and may result in your Agent crashing and/or
creating processes that remain in a stuck or zombie state.

However, they provide a simple helper that accomplishes everything we need
here.

`AgentCheck` import was updated to use the version supported and suggested for
agent version 6 and higher.

[1]: https://github.com/DataDog/datadog-agent/blob/c3d08053dcd44d18e52b27f7562d5d2a78fd84cb/docs/dev/checks/python/check_api.md#running-subprocesses